Add files via upload
There is a take home for SWIFT TRAVELS COMPANY to tighten their belt and watch out for the months people travel most and not be worried about the months they travel less. Rather, they should check the means of transportation they want and channel their energy to it. Kindly check the analysis for more insight from the dashboard.