Add files via upload
This dashboard shows the Human Resources Dashboard. The column chat shows the hierachy by the count of each department where sales and marketing has the highest population with more of men. We can also see the trends of the sum of employees rated in the previous year and the awards won per each department. Operations, Procurement, and Technology Department took the first three and others followed. HR employed more of Bachelor's degree than Master's degree. 

This dashboard is very explicit as more inights can be seen when we play around the dashboard.